#50 Docs: 폴더 변경
